dwarf: fix unaligned access

+++ bld/dwarf/dw/c/dwmac.c
@@ -2,7 +2,7 @@
 *
 *                            Open Watcom Project
 *
-* Copyright (c) 2002-2024 The Open Watcom Contributors. All Rights Reserved.
+* Copyright (c) 2002-2026 The Open Watcom Contributors. All Rights Reserved.
 *    Portions Copyright (c) 1983-2002 Sybase, Inc. All Rights Reserved.
 *
 *  ========================================================================
@@ -47,7 +47,7 @@ typedef struct dw_parm {
 struct dw_macro {
     dw_parm             *parms;
     dw_linenum          line;
-    size_t              len;
+    unsigned            len;
     char                name[1];        /* not nul terminated */
 };
